+++ truffle/com.oracle.truffle.object/src/com/oracle/truffle/object/debug/ShapeProfiler.java
@@ -35,6 +35,7 @@ import java.util.IdentityHashMap;
 import java.util.List;
 import java.util.concurrent.ConcurrentLinkedQueue;
 
+@Deprecated
 public class ShapeProfiler {
     private static final String LINE_SEPARATOR = "***********************************************";
     private static final String BULLET = "* ";